requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.27
sqlalchemy>=2.0
psycopg2-binary>=2.9
pydantic>=2.6
pydantic-settings>=2.3
python-dotenv>=1.0
pika>=1.3
httpx>=0.27
trafilatura>=1.9
beautifulsoup4>=4.12
tiktoken>=0.7
sentence-transformers>=3.0
faiss-cpu>=1.7.4
numpy>=1.26
scikit-learn>=1.4
filelock>=3.13
google-genai>=1.45.0
tenacity>=8.3